# Format requests/responses ahead of time in file audit